gl-renderer: Always setup gl-renderer
Currently, the gl-renderer setup is being done on per-output
basis. This isn't desirable when trying to make weston run
with zero outputs.

When there are no outputs present, there is no EGLContext or
EGLCurrent available, and trying to use an EGL extension will
result in a crash.

The problem is solved by creating a dummy surface at startup,
either by using EGL_KHR_surfaceless_context or PbufferSurface
to setup a context, and creating an EGLContext and EGLCurrent
with the help of dummy surface.

When an output gets plugged in, its configuration will replace
dummy buffer usage as the setup will be run again.

Signed-off-by: Armin Krezović <krezovic.armin@gmail.com>