fullscreen-shell: Remap surface when attaching primary output
When a surface gets created and no outputs are present,
it won't get properly configured. This happens because
call to configure_presented_surface() depends on a call
to fs_output_set_surface() which requires a valid output.

This patch makes the code save the surface when no
outputs are present and reconfigure it when an output
gets attached, so it gets properly displayed.

Signed-off-by: Armin Krezović <krezovic.armin@gmail.com>