compositor: Mark all views as dirty when all outputs are gone
When all outputs are gone and views were created before they
were gone, all views would have an output that points to a
destroyed object.

Instead, mark the view as dirty and set the view output to
NULL so a view gets an output assigned as soon as it gets
plugged in.

Signed-off-by: Armin Krezović <krezovic.armin@gmail.com>