fullscreen-shell: Remap surfaces when attaching a primary output
When no outputs are present, and no output resource was given,
a fullscreen surface won't get configured because the default
behavior when there is no output resource given is to create
the same surface on all fullscreen outputs.

This patch makes the shell reconfigure the surface as soon
as an output gets plugged in and configured, so it can properly
be displayed if it was started when no outputs were present.

Signed-off-by: Armin Krezović <krezovic.armin@gmail.com>