weston: Port X11 backend to pending output API
This is a complete port of the X11 backend that
uses the pending output API and weston_output_api
for output configuration.

- Output can be configured at runtime by passing
  the output (obtained from the pending output
  handler) and the necessary configuration, which
  can be filled in manually or obtained from the
  configuration file by using previously added
  functionality.

- Output can be created at runtime using the
  output API. The output creation only creates
  a pending output, which needs to be configured
  the same way as mentioned above.

Same as before, a single output is created at runtime
using the default configuration or a configuration
from a config file. The output-count functionality
is also preserved, which means more than one output
can be created initially, and more outputs can be
added at runtime using the weston_output_api.

Signed-off-by: Armin Krezović <krezovic.armin@gmail.com>