weston: Port fbdev backend to pending output API
This is a complete port of the fbdev backend that
uses the pending output API and weston_output_api
for output configuration.

Single fbdev output can be configured at runtime
by passing the output (obtained from the pending
output handler) and transform value (as the
output transform is the only thing that can be
configured at the moment).

Same as before, a single framebuffer is opened
at runtime, and the compositor will shutdown if
it can't open it. If it can open it, the pending
output is created and will be configured as
described above.

Signed-off-by: Armin Krezović <krezovic.armin@gmail.com>