weston: Port DRM backend to pending output API
This is a complete port of the DRM backend that
uses the pending output API and weston_output_api
for output configuration.

Outputs can be configured at runtime by passing
the output (obtained from the pending output
handler) and the necessary configuration, which
can be filled in manually or obtained from the
configuration file.

Signed-off-by: Armin Krezović <krezovic.armin@gmail.com>