libweston: Add pending output API
This patch adds a bunch of new functions to libweston
which will allow the compositor to configure an output
after it has been plugged in. All the backends need to
be ported first and the compositor backend configuration
code needs to be reworked for these functions to have
any uses.

I haven't written any documentation on this yet, but I'll
sumarize what each of the functions does, and we can turn
it into documentation once it's agreed upon.

The main function, weston_output_init_pending is responsible
for creating a weston_output object with just enough
information that compositor can look for its configuration
and configure it accordingly (this is a different topic and
will be left for later patch description). It also sets
some (in)sane values for some fields that are expected to
be set to something sane by the backends.

weston_output_set_{scale,transform}, as their name says,
set the output scale and transform. They can be used by
the compositor to configure scale or transform for an
uninitialized output (which has yet to be enabled), and
they can only be used once and not on an output that's
already initialized and enabled.

That leads us to weston_output_enable, which is a result
of combining functionality used by all backends into one
single function. This function is intended to be used
by compositors to indicate it has done setting up the
output and that it should be enabled. It requires calling
to backend specific code, which means all the backends
need updating. So far, it calls weston_output_init,
backend specific code and weston_compositor_add_output,
as done in all the backends. Note that this function
should replace current weston_output_init (which will
be merged later into this one to avoid breaking current
code), and that weston_pending_output_init will become
a new init.

To integrate output configuration into compositor,
weston_pending_output_set_listener function has been
written. It is responsible for setting a callback function
which will be called as soon as new output gets plugged
in and awaits configuration. The function is compositor
specific.

Last function, weston_pending_output_coldplug was added
so compositor can request pending output events to be
re-triggered in case any got emited before the listener
was set up. It is intended to be used as a one-shot
during compositor initialization.

Signed-off-by: Armin Krezović <krezovic.armin@gmail.com>