compositor: Add pending output handling
This adds necessary code for handling the pending outputs.

New wet_compositor fields were introduced so the data
can be passed between different pending outputs (such as
configuration parsed from the command line), as well as
the handler to clear the user data on shutdown, such as
freeing the memory.

Signed-off-by: Armin Krezović <krezovic.armin@gmail.com>