weston: Port DRM backend to pending output API
This is a complete port of the DRM backend that uses
the pending output API and weston_drm_output_api for
output configuration.

Outputs can be configured at runtime by passing the
output (obtained from the pending output handler)
and the necessary configuration, which can be filled
in manually or obtained from the configuration file.

Scale and transform cannot be set using this API
and need to be set explicitly, using previously
added functionality. Output needs to be enabled
manually after it has been fully configured.

Signed-off-by: Armin Krezović <krezovic.armin@gmail.com>