compositor: Add pending output handling
This patch adds new wet_compositor fields for handling the
pending outputs, as well as helpers for accessing those
fields.

This includes a listener for a pending output, whose
handler needs to be set by the by the load_*_backend()
function, as the function is backend-specific.

The optional backend_data field can be used to store
backend specific configuration data, such as options
parsed from the command line. The backend_data_handler
can also be defined if such data need special handling,
such as freeing the memory it has been using.

Signed-off-by: Armin Krezović <krezovic.armin@gmail.com>