weston: Port fbdev backend to new output handling API
This is a complete port of the fbdev backend that uses
recently added output handling API for output
configuration.

It is required that the scale and transform values are
set using the previously added functionality.

After everything has been set, output needs to be
enabled manually using weston_output_enable().

Signed-off-by: Armin Krezović <krezovic.armin@gmail.com>