compositor: Add internal output object used for layout configuration
This adds weston specific output object, which contains information
about output's position, relative to other outputs. DRM and Windowed
backends code has been updated to make use of the new code and
parse new configuration options for a given output, if it was
specified.

New configuration file options for [output] section have been added:

left-of: Specifies which output should be on the right side of the
current output.

right-of: Specifies which output should be on the left side of the
current output.

Signed-off-by: Armin Krezović <krezovic.armin@gmail.com>