compositor: Implement horizontal output layout configuration
This patch adds horizontal output layout configuration using
the previously added code.

When an output is added, it looks for outputs that it should
be placed next to, if they are specified. If such output is
found, output layout is updated if needed (outputs on a given
position are moved right if they collide with the new output)
and the output is positioned next to the specified output,
either on its left or its right.

If a certain position wasn't specified for the current output,
the compositor looks for any existing outputs that have
current output specified on that position, and updates the
positions accordingly, if corresponding output is found.

Output positions can only be set once, where config file
specified option is always set, even if such output will
never be present. If a certain position wasn't set, the
compositor may update the position if it finds another
output that wants to be placed next to the output that's
being configured. In such case, first existing output that
matches the position is picked up, rest is ignored.

Default behaviour is still to place outputs that have
no matching configuration top right.

Signed-off-by: Armin Krezović <krezovic.armin@gmail.com>